master

+++ modules/document/document.controller.php
@@ -268,7 +268,7 @@ class documentController extends document
 		}
 		// If the tile is empty, extract string from the contents.
 		settype($obj->title, "string");
-		if($obj->title == '') $obj->title = cut_str(strip_tags($obj->content),20,'...');
+		if($obj->title == '') $obj->title = cut_str(trim(strip_tags(nl2br($obj->content))),20,'...');
 		// If no tile extracted from the contents, leave it untitled.
 		if($obj->title == '') $obj->title = 'Untitled';
 		// Remove XE's own tags from the contents.
@@ -333,7 +333,10 @@ class documentController extends document
 		$oDB->commit();
 
 		// return
-		$this->addGrant($obj->document_srl);
+		if(!$manual_inserted)
+		{
+			$this->addGrant($obj->document_srl);
+		}
 		$output->add('document_srl',$obj->document_srl);
 		$output->add('category_srl',$obj->category_srl);
 
@@ -440,7 +443,7 @@ class documentController extends document
 		if(Context::get('is_logged'))
 		{
 			$logged_info = Context::get('logged_info');
-			if($source_obj->get('member_srl')==$logged_info->member_srl || $bUseHistory)
+			if($source_obj->get('member_srl')==$logged_info->member_srl)
 			{
 				$obj->member_srl = $logged_info->member_srl;
 				$obj->user_name = htmlspecialchars_decode($logged_info->user_name);
@@ -512,30 +515,33 @@ class documentController extends document
 			return $output;
 		}
 		// Remove all extra variables
-		$this->deleteDocumentExtraVars($source_obj->get('module_srl'), $obj->document_srl, null, Context::getLangType());
-		// Insert extra variables if the document successfully inserted.
-		$extra_keys = $oDocumentModel->getExtraKeys($obj->module_srl);
-		if(count($extra_keys))
+		if(Context::get('act')!='procFileDelete')
 		{
-			foreach($extra_keys as $idx => $extra_item)
+			$this->deleteDocumentExtraVars($source_obj->get('module_srl'), $obj->document_srl, null, Context::getLangType());
+			// Insert extra variables if the document successfully inserted.
+			$extra_keys = $oDocumentModel->getExtraKeys($obj->module_srl);
+			if(count($extra_keys))
 			{
-				$value = NULL;
-				if(isset($obj->{'extra_vars'.$idx}))
+				foreach($extra_keys as $idx => $extra_item)
 				{
-					$tmp = $obj->{'extra_vars'.$idx};
-					if(is_array($tmp))
-						$value = implode('|@|', $tmp);
-					else
-						$value = trim($tmp);
+					$value = NULL;
+					if(isset($obj->{'extra_vars'.$idx}))
+					{
+						$tmp = $obj->{'extra_vars'.$idx};
+						if(is_array($tmp))
+							$value = implode('|@|', $tmp);
+						else
+							$value = trim($tmp);
+					}
+					else if(isset($obj->{$extra_item->name})) $value = trim($obj->{$extra_item->name});
+					if($value == NULL) continue;
+					$this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, $idx, $value, $extra_item->eid);
 				}
-				else if(isset($obj->{$extra_item->name})) $value = trim($obj->{$extra_item->name});
-				if($value == NULL) continue;
-				$this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, $idx, $value, $extra_item->eid);
 			}
+			// Inert extra vars for multi-language support of title and contents.
+			if($extra_content->title) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -1, $extra_content->title, 'title_'.Context::getLangType());
+			if($extra_content->content) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -2, $extra_content->content, 'content_'.Context::getLangType());
 		}
-		// Inert extra vars for multi-language support of title and contents.
-		if($extra_content->title) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -1, $extra_content->title, 'title_'.Context::getLangType());
-		if($extra_content->content) $this->insertDocumentExtraVar($obj->module_srl, $obj->document_srl, -2, $extra_content->content, 'content_'.Context::getLangType());
 		// Update the category if the category_srl exists.
 		if($source_obj->get('category_srl') != $obj->category_srl || $source_obj->get('module_srl') == $logged_info->member_srl)
 		{
@@ -816,6 +822,9 @@ class documentController extends document
 	 */
 	function updateReadedCount(&$oDocument)
 	{
+		// Pass if Crawler access
+		if(isCrawler()) return false;
+		
 		$document_srl = $oDocument->document_srl;
 		$member_srl = $oDocument->get('member_srl');
 		$logged_info = Context::get('logged_info');
@@ -849,7 +858,7 @@ class documentController extends document
 		$output = executeQuery('document.updateReadedCount', $args);
 
 		// Call a trigger when the read count is updated (after)
-		$trigger_output = ModuleHandler::triggerCall('document.updateReadedCount', 'after', $oDocument);
+		$outptrigger_outputut = ModuleHandler::triggerCall('document.updateReadedCount', 'after', $oDocument);
 		if(!$trigger_output->toBool())
 		{
 			$oDB->rollback();
@@ -867,7 +876,10 @@ class documentController extends document
 		}
 
 		// Register session
-		$_SESSION['readed_document'][$document_srl] = true;
+		if(!$_SESSION['banned_document'][$document_srl]) 
+		{
+			$_SESSION['readed_document'][$document_srl] = true;
+		}
 
 		return TRUE;
 	}
@@ -2473,18 +2485,13 @@ class documentController extends document
 		if(is_array($documentSrlList))
 		{
 			$documentSrlList = array_unique($documentSrlList);
-			foreach($documentSrlList AS $key=>$documentSrl)
+			foreach($documentSrlList AS $key => $documentSrl)
 			{
-				$oldDocument = $oDocumentModel->getDocument($documentSrl);
 				$fileCount = $oFileModel->getFilesCount($documentSrl);
-
-				if($oldDocument != null)
-				{
-					$newDocumentArray = $oldDocument->variables;
-					$newDocumentArray['uploaded_count'] = $fileCount;
-					$newDocumentObject = (object) $newDocumentArray;
-					$this->updateDocument($oldDocument, $newDocumentObject);
-				}
+				$args = new stdClass();
+				$args->document_srl = $documentSrl;
+				$args->uploaded_count = $fileCount;
+				executeQuery('document.updateUploadedCount', $args);
 			}
 		}
 	}
